extract a bunch of graph view functions
